+++ devicehive/__init__.py
@@ -25,7 +25,7 @@ from twisted.internet import reactor
 from urlparse import urlsplit, urljoin
 
 
-__all__ = ['HTTP11DeviceHiveFactory', 'DeviceDelegate', 'Equipment', 'CommandResult']
+__all__ = ['HTTP11DeviceHiveFactory', 'DeviceDelegate', 'Equipment', 'CommandResult', 'BaseHTTP11ClientFactory']
 
 
 def parse_url(device_hive_url) :
@@ -121,7 +121,7 @@ class BaseRequest(Request):
     def default_headers(self, host, device_id, device_key):
         headers = Headers({'Host': [host,],
                             'Content-Type': ['application/json',],
-                               'Auth-DeviceID': [device_id,],
+                            'Auth-DeviceID': [device_id,],
                             'Auth-DeviceKey': [device_key,],
                             'Accept': ['application/json',]})
         return headers
@@ -486,8 +486,8 @@ class DeviceDelegate(object):
         Sends notification to Device-Hive server.
         """
         if self.factory is not None :
-            self.factory.next_state(ProtocolState.Notify, data = NotifyData(notification, kwargs))
-
+            self.factory.notify(notification, kwargs)
+    
     def registration_info(self):
         res = {'id': self.device_id(),
         'key': self.device_key(),
@@ -606,6 +606,9 @@ class BaseHTTP11ClientFactory(ClientFactory):
             self.state.do_retry = True
         if (not self.started) and (connector is not None) and (connector.state == 'disconnected') :
             self.handleConnectionLost(connector)
+    
+    def notify(self, notification, kwargs):
+        raise NotImplementedError()
 
 
 class _SingleRequestHTTP11DeviceHiveFactory(BaseHTTP11ClientFactory):
@@ -625,394 +628,6 @@ class _SingleRequestHTTP11DeviceHiveFactory(BaseHTTP11ClientFactory):
             raise NotImplementedError('Unsupported factory state <{0}>.'.format(self.state.value))
 
 
-class WebSocketError(Exception):
-    def __init__(self, msg = '') :
-        super(WebSocketError, self).__init__('WebSocket error. Reason: {0}.'.format(msg))
-
-
-WS_OPCODE_CONTINUATION = 0
-WS_OPCODE_TEXT_FRAME   = 1
-WS_OPCODE_BINARY_FRAME = 2
-WS_OPCODE_CONNECTION_CLOSE = 8
-WS_OPCODE_PING = 9
-WS_OPCODE_PONG = 10
-
-
-WS_GUID = '258EAFA5-E914-47DA-95CA-C5AB0DC85B11'
-
-
-class WebSocketState(Values) :
-    Status     = ValueConstant(0)
-    Header     = ValueConstant(1)
-    WsHeader   = ValueConstant(2)
-    WsLength7  = ValueConstant(3)
-    WsLength16 = ValueConstant(4)
-    WsLength64 = ValueConstant(5)
-    WsPayload  = ValueConstant(6)
-
-
-class IWebSocketHandler(Interface) :
-    def status_received(self, proto_version, code, status) :
-        pass
-    
-    def header_received(self, name, value):
-        pass
-    
-    def headers_received(self) :
-        pass
-    
-    def frame_received(self, opcode, payload):
-        pass
-
-
-class WebSocketParser(LineReceiver) :
-    """
-    Class parses incoming byte stream and extracts HTTP headers and WebSocket frames.
-    """
-    
-    def __init__(self, handler):
-        self.state = WebSocketState.Status
-        self.handler = handler
-        # attribute to store header
-        self._header_buf = None
-        # attributes which store frame
-        # data and parameters
-        self._frame_buf = b''
-        self._frame_fin = False
-        self._frame_opcode = 0
-        self._frame_len = 0
-        self._frame_data = b''
-    
-    def lineReceived(self, line):
-        if line[-1:] == '\r':
-            line = line[:-1]
-        
-        if self.state == WebSocketState.Status :
-            self.status_received(line)
-            self.state = WebSocketState.Header
-        elif self.state == WebSocketState.Header :
-            if not line or line[0] not in ' \t':
-                if self._header_buf is not None:
-                    header = ''.join(self._header_buf)
-                    name, value = header.split(':', 1)
-                    value = value.strip()
-                    self.header_received(name, value)
-                
-                if not line:
-                    self.headers_received()
-                else:
-                    self._header_buf = [line]
-            else:
-                self._header_buf.append(line)
-    
-    def status_received(self, line) :
-        if (self.handler is not None) and IWebSocketHandler.implementedBy(self.handler.__class__) :
-            proto_version, code, status = line.split(' ', 2)
-            self.handler.status_received(proto_version, int(code, 10), status)
-    
-    def header_received(self, name, value):
-        if (self.handler is not None) and IWebSocketHandler.implementedBy(self.handler.__class__) :
-            self.handler.header_received(name, value)
-    
-    def headers_received(self) :
-        if (self.handler is not None) and IWebSocketHandler.implementedBy(self.handler.__class__) :
-            self.handler.headers_received()
-        self.state = WebSocketState.WsHeader
-        self.setRawMode()
-    
-    def frame_received(self, opcode, payload):
-        if (self.handler is not None) and IWebSocketHandler.implementedBy(self.handler.__class__) :
-            self.handler.frame_received(opcode, payload)
-    
-    def rawDataReceived(self, data):
-        self._frame_buf += data
-        #
-        while True :
-            if self.state == WebSocketState.WsHeader :
-                if len(self._frame_buf) > 0 :
-                    hdr = struct.unpack('B', self._frame_buf[:1])[0]
-                    self._frame_buf = self._frame_buf[1:]
-                    self._frame_fin = (hdr & 0x80)
-                    self._frame_opcode = (hdr & 0x0f)
-                    self.state = WebSocketState.WsLength7
-                else :
-                    break
-            elif self.state == WebSocketState.WsLength7 :
-                if len(self._frame_buf) > 0 :
-                    len7 = struct.unpack('B', self._frame_buf[:1])[0]
-                    self._frame_buf = self._frame_buf[1:]
-                    if len7 & 0x80 :
-                        raise WebSocketError('Server should not mask websocket frames.')
-                    else :
-                        len7 = len7 & 0x7f
-                        if len7 == 126 :
-                            self._frame_len = 0
-                            self.state = WebSocketState.WsLength16
-                        elif len7 == 127 :
-                            self._frame_len = 0
-                            self.state = WebSocketState.WsLength64
-                        else :
-                            self._frame_len = len7
-                            self.state = WebSocketState.WsPayload
-                else :
-                    break
-            elif self.state == WebSocketState.WsLength16 :
-                if len(self._frame_buf) > 1 :
-                    len16 = struct.unpack('!H', self._frame_buf[:2])[0]
-                    self._frame_buf = self._frame_buf[2:]
-                    self._frame_len = len16
-                    self.state = WebSocketState.WsPayload
-                else :
-                    break
-            elif self.state == WebSocketState.WsLength64 :
-                if len(self._frame_buf) > 7 :
-                    len64 = struct.unpack('!Q', self._frame_buf[:8])[0]
-                    self._frame_buf = self._frame_buf[8:]
-                    self._frame_len = len64
-                    self.state = WebSocketState.WsPayload
-                else :
-                    break
-            elif self.state == WebSocketState.WsPayload :
-                if self._frame_len == 0 :
-                    if self._frame_fin :
-                        self.frame_received(self._frame_opcode, self._frame_data)
-                        self._frame_data = b''
-                        self._frame_opcode = 0
-                    self.state = WebSocketState.WsHeader
-                elif len(self._frame_buf) == 0 :
-                    break
-                else :
-                    bytes_to_read = min(self._frame_len, len(self._frame_buf))
-                    self._frame_data += self._frame_buf[:bytes_to_read]
-                    self._frame_buf = self._frame_buf[bytes_to_read:]
-                    self._frame_len -= bytes_to_read
-            elif self.state == WebSocketState.WsError :
-                break
-        pass
-
-
-class WebSocketProtocol13(Protocol):
-    implements(IWebSocketHandler)
-    
-    def __init__(self, factory):
-        self.parser  = WebSocketParser(self)
-        self.factory = factory
-        self.security_key = base64.b64encode((uuid.uuid4()).bytes)
-        self.handshaked = False
-        self.rand = Random(long(time()))
-        self.mask = b'\x00\x00\x00\x00'
-    
-    def connectionLost(self, reason):
-        pass
-    
-    def connectionMade(self):
-        pass
-    
-    def dataReceived(self, data):
-        self.parser.dataReceived(data)    
-    
-    def status_received(self, proto_version, code, status) :
-        if proto_version != 'HTTP/1.1' :
-            raise WebSocketError('unsupported protocol {0}'.format(proto_version))
-        if code != 101 :
-            raise WebSocketError('websocket server rejected protocol upgrade with code {0}'.format(code))
-    
-    def header_received(self, name, value):
-        """
-        Verifies headers on the fly. If there is an error connection would be aborted.
-        """
-        loname = name.lower()
-        if loname == 'sec-websocket-accept' :
-            if not self.validate_security_answer(value) :
-                raise WebSocketError('websocket server returned invalid security key {0} in response to {1}'.format(value, self.security_key))
-        elif loname == 'connection' :
-            if value.lower() != 'upgrade' :
-                raise WebSocketError('websocket server failed to upgrade connection, status = {0}'.format(value))
-        elif loname == 'upgrade' :
-            if value.lower() != 'websocket' :
-                raise WebSocketError('websocket server upgraded protocol to invalid state {0}'.format(value))
-    
-    def headers_received(self) :
-        pass
-    
-    def frame_received(self, opcode, payload):
-        if opcode == WS_OPCODE_PING :
-            self.send_frame(1, WS_OPCODE_PONG, payload)
-        elif opcode == WS_OPCODE_PONG :
-            pass # do nothing
-        else :
-            raise NotImplementedError()
-    
-    def validate_security_answer(self, answer):
-        skey = sha.new(self.security_key + WS_GUID)
-        key = base64.b64encode(skey.digest())
-        return answer == key
-    
-    def send_headers(self) :
-        header = 'GET /device HTTP/1.1\r\n' + \
-                  'Host: {0}\r\n' + \
-                  'Auth-DeviceID: {1}\r\n' + \
-                  'Auth-DeviceKey: {2}\r\n' + \
-                  'Upgrade: websocket\r\n' + \
-                  'Connection: Upgrade\r\n' + \
-                  'Sec-WebSocket-Key: {3}' + \
-                  'Origin: http://{0}\r\n' + \
-                  'Sec-WebSocket-Protocol: device-hive, devicehive\r\n' + \
-                  'Sec-WebSocket-Version: 13\r\n\r\n'
-        return header.format(self.factory.host,
-                             self.factory.device_delegate.device_id(),
-                             self.factory.device_delegate.device_key(),
-                             self.security_key).encode('utf-8')
-    
-    def send_frame(self, fin, opcode, data) :
-        frame = struct.pack('B', (0x80 if fin else 0x00) | opcode)[0]
-        l = len(data)
-        if l < 126:
-            frame += struct.pack('B', l | 0x80)[0]
-        elif l <= 0xFFFF:
-            frame += struct.pack('!BH', 126 | 0x80, l)[0]
-        else:
-            frame += struct.pack('!BQ', 127 | 0x80, l)[0]
-        self.mask  = chr(self.rand.randint(0, 0xff)) + chr(self.rand.randint(0, 0xff)) + chr(self.rand.randint(0, 0xff)) + chr(self.rand.randint(0, 0xff))
-        frame += self.mask
-        frame += array('B', [ord(data[i]) ^ ord(self.mask[i % 4]) for i in range(len(data))]).tostring()
-        self.transport.write(frame)
-
-
-class WebSocketDeviceHiveProtocol(HTTP11ClientProtocol):
-    """
-    Implements Device-Hive protocol over WebSockets
-    """
-    
-    WS_GUID = '258EAFA5-E914-47DA-95CA-C5AB0DC85B11'
-    
-    def request_counter() :
-        request_number = 1
-        while True :
-            yield request_number
-            request_number += 1
-    request_counter = request_counter()
-    
-    def __init__(self, factory) :
-        if hasattr(HTTP11ClientProtocol, '__init__'):
-            HTTP11ClientProtocol.__init__(self)
-        self.factory = factory
-        self.parser = WebSocketParser(self)
-        self.security_key = base64.b64encode((uuid.uuid4()).bytes)
-    
-    def connectionMade(self) :
-        if self.factory.state.value == ProtocolState.ApiMetadata :
-            res = self.request(ApiMetadataRequest(self.factory))
-            res.addCallbacks(self._apimetadata_done, self._critical_error)
-        elif self.factory.state.value == ProtocolState.Register :
-            self.factory.state = StateHolder(ProtocolState.Registering, self.state.data, self.state.retries)
-            self.send_headers()
-        else :
-            log.err('Unsupported WebSocket API state {0}.'.format( self.factory.state))    
-    
-    def send_headers(self) :
-        self.transport.write(self.default_headers())
-        self.transport.write('\r\n'.encode('utf-8'))
-    
-    def authenticate(self) :
-        """
-        Sends authentication information to WebSocket server.
-        """
-        auth_request = json.dumps({'action': 'authenticate',
-                        'requestId': WebSocketDeviceHiveProtocol.request_counter.next(),
-                        'deviceId': self.factory.device_delegate.device_id(),
-                        'deviceKey': self.factory.device_delegate.device_key()})
-        self.transport.write(auth_request)
-    
-    def status_received(self, proto_version, code, result) :
-        if proto_version != 'HTTP/1.1' :
-            log.err('Terminating WebSocket protocol. Reason: server returned unsupported protocol {0}.'.format(proto_version))
-        elif code != 101 :
-            log.err('Terminating WebSocket protocol. Reason: server returned status code {0}.'.format(code))
-    
-    def headers_received(self) :
-        """
-        All headers have been received.
-        """
-        self.authenticate()
-    
-    def dataReceived(self, bytes) :
-        if self.factory.state.value == ProtocolState.ApiMetadata :
-            HTTP11ClientProtocol.dataReceived(self, bytes)
-        else :
-            self.parser.dataReceived(bytes)
-    
-    
-    def _apimetadata_done(self, response) :
-        log.msg('ApiInfo respond: {0}.'.format( response ))
-        if response.code == 200 :
-            def get_response(resp, factory, connector) :
-                log.msg('ApiInfo response {0} has been successfully received.'.format(resp))
-                if hasattr(factory, 'on_apiinfo_finished') and callable(factory.on_apiinfo_finished) :
-                    factory.on_apiinfo_finished(resp, connector)
-            
-            def err_response(reason, connector) :
-                log.msg('Failed to receive ApiInfo response. Reason: {0}.'.format(reason))
-                self.factory.retry(connector)
-            
-            result_proto = Deferred()
-            result_proto.addCallbacks(partial(get_response, factory = self.factory, connector = self.transport.connector), partial(err_response, connector = self.transport.connector))
-            response.deliverBody(JsonDataConsumer(result_proto))
-        else :
-            def get_response_text(reason):
-                log.err('ApiInfo call failed. Response: <{0}>. Code <{1}>. Reason: <{2}>.'.format(response, response.code, reason))
-            response_defer = Deferred()
-            response_defer.addCallbacks(get_response_text, get_response_text)
-            response.deliverBody(TextDataConsumer(response_defer))
-            self.factory.retry(self.transport.connector)
-    
-    def _critical_error(self, reason) :
-        log.err("Device-hive websocket api failure. Critical error: <{0}>".format(reason))
-        if reactor.running :
-            if hasattr(self.factory, 'on_failure') and callable(self.factory.on_failure) :
-                self.factory.on_failure()
-
-
-class WebSocketDeviceHiveFactory(BaseHTTP11ClientFactory):
-    def __init__(self, device_delegate, retries = 3):
-        BaseHTTP11ClientFactory.__init__(self, StateHolder(ProtocolState.Unknown), retries)
-        self.uri = 'localhost'
-        self.host = 'localhost'
-        self.port = 80
-        self.device_delegate = device_delegate
-        self.device_delegate.factory = self
-        self.server_timestamp = None
-    
-    def doStart(self):
-        if self.state.value == ProtocolState.Unknown :
-            self.state = StateHolder(ProtocolState.ApiMetadata)
-        BaseHTTP11ClientFactory.doStart(self)
-    
-    def buildProtocol(self, addr):
-        return WebSocketDeviceHiveProtocol(self)
-    
-    def handleConnectionLost(self, connector) :
-        if self.state.value == ProtocolState.Register :
-            log.msg('Connecting to WebSocket server {0}:{1}.'.format(self.host, self.port))
-            reactor.connectTCP(self.host, self.port, self)
-        else :
-            log.msg('Quiting WebSocket factory.')
-    
-    def on_apiinfo_finished(self, response, connector) :
-        self.uri, self.host, self.port = parse_url(response['webSocketServerUrl'])
-        log.msg('WebSocket service location: {0}, Host: {1}, Port: {2}.'.format(self.uri, self.host, self.port))
-        #
-        self.uri = response['webSocketServerUrl']
-        self.server_timestamp = parse_date(response['serverTimestamp'])
-        self.state = StateHolder(ProtocolState.Register)
-    
-    def on_registration_finished(self, reason=None):
-        log.msg('Registration finished for reason: {0}.'.format(reason))
-    
-    def on_failure(self):
-        log.msg('On failure')
-
-
 class HTTP11DeviceHiveFactory(BaseHTTP11ClientFactory):
     """
     L{DeviceHiveFactory} is an implementation of the DeviceHive protocol v6.
@@ -1028,7 +643,7 @@ class HTTP11DeviceHiveFactory(BaseHTTP11ClientFactory):
     """
     def __init__(self, device_delegate, poll_interval = 1.0, retries = 3):
         BaseHTTP11ClientFactory.__init__(self, StateHolder(ProtocolState.Unknown), retries)
-        self.uri  = 'localhost'
+        self.uri = 'localhost'
         self.host = 'localhost'
         self.port = 80
         self.poll_interval = poll_interval
@@ -1061,7 +676,7 @@ class HTTP11DeviceHiveFactory(BaseHTTP11ClientFactory):
                 while tmp_state.value == ProtocolState.Notify :
                     self.single_request(tmp_state)
                     tmp_state = None
-                    if len(self.states_stack) > 0:
+                    if len(self.states_stack) > 0 :
                         tmp_state = self.states_stack.pop(0)
                     else :
                         return
@@ -1069,8 +684,10 @@ class HTTP11DeviceHiveFactory(BaseHTTP11ClientFactory):
                 if tmp_state is not None :
                     self.state = tmp_state
                     reactor.callLater(self.poll_interval, reconnect, connector)
-                #
         pass
+    
+    def notify(self, notification, kwargs) :
+        self.next_state(ProtocolState.Notify, data = NotifyData(notification, kwargs))
     
     def next_state(self, next_state, data = None, connector = None) :
         if self.registered and (next_state == ProtocolState.Notify or next_state == ProtocolState.Report) :
